Fix Products Details. Code refactor Products Route and Controller
