+++ src/views/products/index.ejs
@@ -1,0 +1,1 @@
+<h1>Productos</h1>
