--- src/views/products/productDetail.ejs
@@ -1,104 +1,0 @@
-<%- include('../partials/metadata.ejs') %> 
-
-<%- include('../partials/header.ejs') %> 
-    <main>
-        <section class="product-main">
-            <article class="product-img">
-                <div>
-                    <img src="\images\alimento-excellent-gatos.png" alt="product-excellent">
-                    <button type="button">
-                        <i class="fa-solid fa-magnifying-glass"></i>
-                    </button>
-                </div>
-            </article>
-            <article class="product-info-main">
-                <div class="product-info">
-                    <h3> Excellent Gato Adulto 7.5kg.</h3>
-                    <p> Referencia Excellent Gato Adulto 7.5kg.</p>
-                    <p class="price-product"> $7.200,00</p>
-                </div>
-                <div class="product-info-chart">
-                    <h4> Excellent Gato Adulto</h4>
-                    <p> Pollo y Arroz, ofrece una fórmula con proteínas
-                        de alta digestibilidad y ácidos grasos omega 3 y 6,
-                        para mantener al máximo la salud del gato adulto.</p>
-                    <br>
-                    <div class="product-chart">
-                        <div class="number-chart">
-                            <input type="number" value="1" min="1">
-                            <button type="button">
-                                <i class="fa fa-angle-up"></i>
-                            </button>
-                            <button type="button">
-                                <i class="fa fa-angle-down"></i>
-                            </button>
-                        </div>                        
-                        <div class="add-chart">
-                            <button type="submit">
-                                <i class="fa-sharp fa-solid fa-cart-plus"></i>
-                                Añadir al carrito
-                            </button>                    
-                            <button type="button" alt="Añadir a lista de deseados">
-                                <i class="fa-solid fa-heart"></i>
-                            </button>
-                            <button type="button" alt="Comparar producto">
-                                <i class="fa-solid fa-shuffle"></i>
-                            </button>
-                        </div>                        
-                    </div>
-                </div>
-                <br>
-                <div class="product-payment">
-                    <i class="fa-sharp fa-solid fa-credit-card"></i>
-                    <p> ¡PAGOS A TRAVÉS DE MERCADO PAGO
-                        & POSIBILIDAD DE CUOTAS!</p>
-                    <img src="\images\LOGOS-DE-BANCOS.png" alt="LOGOS-DE-BANCOS">
-                </div>
-                <br>
-                <div class="product-shipping">
-                    <p>POR CONSULTAS O DISP. DE STOCK NO DUDE
-                        EN CONTACTARNOS A NUESTRO WHATSAPP</p>
-                    <br>
-                    <i class="fa-solid fa-truck"></i>
-                    <p>ENVÍOS SIN CARGO A C.A.B.A & ZONA OESTE
-
-                        (resto del país a cargo del comprador)</p>
-                    <br>
-                    <i class="fa-regular fa-credit-card"></i>
-                    <p>Compra mínima $1.500</p>
-                </div>
-                <div class="product-socialMedia">
-                    <a href="https://www.whatsapp.com"><img class="logoRedes" src="/images/logoWhatsapp.png"
-                            alt="logo wtsp"></a>
-                    <p>Whatsapp</p>
-                    <a href="https://www.instagram.com"><img class="logoRedes" src="/images/logoIg.png"
-                            alt="logo ig"></a>
-                    <p>Instagram</p>
-                    <a href="https://www.twitter.com"><img class="logoRedes" src="/images/logoTwtr.png"
-                            alt="logo twtr"></a>
-                    <p>Twitter</p>
-                    <a href="https://www.facebook.com"><img class="logoRedes" src="/images/logoFace.png"
-                            alt="logo fb"></a>
-                    <p>Facebook</p>
-                </div>
-            </article>
-        </section>
-        <section class="product-description">
-            <br>
-            <h4>Descripcion del producto</h4>
-            <p>La utilización de fuentes de proteínas de elevada digestibilidad garantiza el aporte de aminoácidos
-                esenciales que ayudan a mantener músculos fuertes y sanos. Así, cualquiera sea el nivel de actividad
-                física de tu gato, puedes estar seguro de que sus necesidades proteicas estarán
-                satisfechas.</p>
-
-            <p>Ayuda a mantener un pH adecuado en la orina, el cual contribuye a la reducción de la formación de
-                cálculos y cristales, colaborando con el mantenimiento de un tracto urinario saludable, fundamental
-                para la salud de tu gato.</p>    
-
-            <p>Piel sana y pelaje brillante. La combinación de los ácidos grasos Omega 6 y 3 junto a vitaminas y
-                minerales favorecen la salud de la piel y un pelaje brillante, primera señal de una salud general
-                adecuada.</p> 
-            <br>
-        </section>
-    </main>
-    <%- include('../partials/footer.ejs') %>
